Moved a11y testing to devsecops folder, replaced timestamp with sha for file name
