+++ devsecops/axe-testing/src/check-bilingual-aria-labels.js
@@ -12,23 +12,32 @@ export async function detectAriaBilingualIssues(page) {
     );
   });
 
+  // console.log('Extracted ARIA Labels:', ariaLabels); // Check extracted labels
+
+
   const ariaBilingualIssues = ariaLabels
     .map(({ html, label }) => {
-      const detectedLanguage = franc(label, { minLength: 3 });
+      // const detectedLanguage = franc(label, { minLength: 3 });
+      const detectedLanguage = franc(label || '', { minLength: 3 });
 
       const containsFrench = detectedLanguage === 'fra';
       const containsEnglish = detectedLanguage === 'eng';
 
+      // console.log(`Label: "${label}", Detected Language: ${detectedLanguage}`);
+
+
       if (!containsFrench || !containsEnglish) {
         return {
-          html,
           label,
-          message:
-            'ARIA label does not contain both French and English content',
+          detectedLanguage,
+          html,
+          message: 'ARIA label does not contain both French and English content',
         };
       }
       return null;
     })
     .filter((issue) => issue !== null);
+  
+  // console.log('Bilingual Issues Detected:', ariaBilingualIssues); 
   return ariaBilingualIssues;
 }
